fixed similar bug in rar postprocessing

+++ pynab/rars.py
@@ -20,36 +20,37 @@ def check_release_files(server, group_name, nzb):
 
     rar_files = []
     for rar in nzb['rars']:
-        log.debug(pprint.pformat(rar))
-
-        if not isinstance(rar['segments'], list):
-            rar['segments']['segment'] = [rar['segments']['segment'], ]
+        pprint.pprint(rar)
         messages = []
+        if not isinstance(rar['segments']['segment'], list):
+            rar['segments']['segment'] = [rar['segments']['segment'], ]
         for s in rar['segments']['segment']:
-            for p in s:
-                messages.append(p['#text'])
+            messages.append(s['#text'])
 
-        data = server.get(group_name, messages)
+        if messages:
+            data = server.get(group_name, messages)
 
-        if data:
-            t = None
-            try:
-                with tempfile.NamedTemporaryFile('wb', delete=False) as t:
-                    t.write(data.encode('ISO-8859-1'))
-                    t.flush()
-                rar_files += lib.rar.RarFile(t.name).infolist()
-            except:
-                continue
-            finally:
-                log.debug('Deleting temporary file {}...'.format(t.name))
-                os.remove(t.name)
-            break
+            if data:
+                t = None
+                try:
+                    with tempfile.NamedTemporaryFile('wb', delete=False) as t:
+                        t.write(data.encode('ISO-8859-1'))
+                        t.flush()
+                    rar_files += lib.rar.RarFile(t.name).infolist()
+                except:
+                    continue
+                finally:
+                    log.debug('Deleting temporary file {}...'.format(t.name))
+                    os.remove(t.name)
+                break
 
-    passworded = any([r.is_encrypted for r in rar_files])
-    file_count = len(rar_files)
-    size = sum([r.file_size for r in rar_files])
+            passworded = any([r.is_encrypted for r in rar_files])
+            file_count = len(rar_files)
+            size = sum([r.file_size for r in rar_files])
 
-    return (passworded, file_count, size), rar_files
+            return (passworded, file_count, size), rar_files
+
+    return (False, 0, 0), []
 
 
 def process(limit=20):
@@ -61,27 +62,37 @@ def process(limit=20):
         for release in db.releases.find({'passworded': None}).limit(limit):
             nzb = pynab.nzbs.get_nzb_dict(release['nzb'])
 
-            (passworded, file_count, size), rar_files = check_release_files(server, release['group']['name'], nzb)
-            if not passworded:
-                for file in rar_files:
-                    if PASSWORDED_REGEX.search(file.filename):
-                        log.debug('Release definitely passworded.')
-                        passworded = True
-                    elif MAYBE_PASSWORDED_REGEX.search(file.filename):
-                        log.debug('Release potentially passworded.')
-                        passworded = config.site['delete_potentially_passworded']
+            if nzb['rars']:
+                (passworded, file_count, size), rar_files = check_release_files(server, release['group']['name'], nzb)
+                if not passworded:
+                    for file in rar_files:
+                        if PASSWORDED_REGEX.search(file.filename):
+                            log.debug('Release definitely passworded.')
+                            passworded = True
+                        elif MAYBE_PASSWORDED_REGEX.search(file.filename):
+                            log.debug('Release potentially passworded.')
+                            passworded = config.site['delete_potentially_passworded']
+                else:
+                    log.debug('RAR was encrypted.')
+
+                log.info('Adding file data to release: {}'.format(release['name']))
+
+                db.releases.update({'_id': release['_id']}, {
+                    '$set': {
+                        'file_count': file_count,
+                        'size': size,
+                        'passworded': passworded
+                    }
+                })
             else:
-                log.debug('RAR was encrypted.')
-
-            log.info('Adding file data to release: {}'.format(release['name']))
-
-            db.releases.update({'_id': release['_id']}, {
-                '$set': {
-                    'file_count': file_count,
-                    'size': size,
-                    'passworded': passworded
-                }
-            })
+                log.debug('No RARs in release, blacklisting...')
+                db.releases.update({'_id': release['_id']}, {
+                    '$set': {
+                        'file_count': 0,
+                        'size': 0,
+                        'passworded': False
+                    }
+                })
 
     if config.site['delete_passworded']:
         log.info('Deleting passworded releases...')
